+++ lib/mix/test/test_helper.exs
@@ -75,7 +75,7 @@ defmodule MixTest.Case do
   end
 
   def in_fixture(which, tmp, function) do
-    src  = Path.join fixture_path(which), "."
+    src  = fixture_path(which) <> "/"
     dest = tmp_path(tmp)
     flag = tmp_path |> String.to_char_list!